requirements.txt
soundfile
numpy
librosa
datasets
huggingface_hub
python-dotenv
defusedxml
